actualizacion en carrito

+++ assets/js/_funciones.js
@@ -1,8 +1,6 @@
 import { BD_Productos, BD_Tiendas } from '../../assets/js/_baseDatos.js';
 import { Carrito } from './_clases.js';
 
-let id = 0;
-
 const shop_cards = () => {
 
     const section = document.querySelector(".contact_section__item--bg");
@@ -65,12 +63,12 @@ const verificar_storage = (modal, storage) => {
     const cont_carrito = document.querySelector("#cont-carrito");
     let i = 1;
 
-    if(storage !== null){
+    if(storage !== null && storage !== []){
         
         storage.map( e => {
             cont_carrito.textContent = ` ${i++}`;
 
-            return modal.innerHTML += `
+            modal.innerHTML += `
                 <div class="card mb-3" style="max-width: 540px;">
                     <div class="row g-0">
                         <div class="col-md-4">
@@ -97,8 +95,8 @@ const verificar_storage = (modal, storage) => {
                         </div>
                     </div>
                 </div>
-            `
-            });
+            `;
+        });
     }else{
         const p = document.createElement('p');
         p.setAttribute("style", "color:black");
@@ -111,59 +109,75 @@ const shopCart = () => {
     const compra = new Carrito();
     const productos = document.querySelectorAll(".agregar-carrito");
     const modal_content = document.querySelector(".modal-body");
-    const products_storaged = JSON.parse(localStorage.getItem("carrito"));
+    let products_storaged = JSON.parse(localStorage.getItem("carrito"));
 
     productos.forEach(producto => {
         producto.addEventListener("click", e => {
-            
-            compra.agregarProducto(e, id);
-            id++;
-            
+            modal_content.replaceChildren()
+            compra.agregarProducto(e);
+
+            const products = JSON.parse(localStorage.getItem("carrito"));
+
+            verificar_storage(modal_content, products);
+            pago_total(products);
+            delete_products(compra);
         });
     });
-
+    
     verificar_storage(modal_content, products_storaged);
-    delete_products(compra);
     pago_total(products_storaged);
+    delete_products(compra);
 }
 
 const delete_products = (compra) => {
     const del = document.querySelectorAll(".delete");
-    
+    const cont_carrito = document.querySelector("#cont-carrito");
+    let i = -1;
+
     del.forEach(btn => {
+        i++;
         btn.addEventListener("click", (e) => {
             const delete_card = e.target.parentElement.parentElement.parentElement.parentElement.parentElement.parentElement.parentElement.parentElement;
 
             compra.borrarProducto(e);
             delete_card.remove();
+            console.log(i)
+            cont_carrito.textContent = ` ${ i--}`;
         });
     })
     
 }
 
 const pago_total = (products_storaged) => {
+    const cont_carrito = document.querySelector("#cont-carrito");
     const tb_precio = document.querySelector("#precio_total");
     const comprar = document.querySelector("#comprar");
+    const body_cart = document.querySelector('.modal-body');
     let precio_total = 0;
 
-    products_storaged.forEach(element => {
-        precio_total += element.precio;
-    });
-    tb_precio.textContent += ` $${precio_total}`;
-
-    comprar.addEventListener("click", () => {
-        const body_cart = document.querySelector('.modal-body');
-
-        alert(
-        `
-            Compra exitosa! 
-            Nos contactaremos vía mail para informarte del envío.
-            Muchas gracias por elegir Mala Influencia Ind.
-        `);
-
-        body_cart.remove();
-        localStorage.clear();
-    });
+    if (products_storaged === null){
+        tb_precio.textContent = `Total: $${precio_total}`;
+    }else{
+        products_storaged.forEach(element => {
+            precio_total += element.precio;
+        });
+        tb_precio.textContent = `Total: $${precio_total}`;
+    
+        comprar.addEventListener("click", () => {
+    
+            alert(
+            `
+                Compra exitosa! 
+                Nos contactaremos vía mail para informarte del envío.
+                Muchas gracias por elegir Mala Influencia Ind.
+            `);
+    
+            body_cart.remove();
+            localStorage.clear();
+            cont_carrito.textContent = "";
+            tb_precio.textContent = `Total: $0`;
+        });  
+    }
 }
 
 export {
